dave: fix strange freefall bug + cleanup
When freefalling to the side from a brick on level0 dave should not
land on the diagonal brick beneath it.
After fixing this, I had to fix the jump cycle as it looked unlike
original.

Overall I think it feels more like the real thing now. Specially when
jumping in tunnles in level1 & level3.