audio: added soundfx module with 'got-trophy' as first sound effect.
It took me some time to research and understand how audio works in original dave.

First it was the format and the inverse-frequency nature of it, however this had some documentation online.
Once figured out how to decode the format, I could produce audio which is similar to what I would hear
when running the original game, but it had different 'pitch' to it.

After a day of research I found out I was missing the frequency stitching in 'zero-crossing'.
Even now the pitch is not perfect as I always finish current frequency symbol to be full instead of
terminating it abruptly. This results little longer sound effects, but right now it feels good enough to continue.