monsters: improve projectile creation to mimic original
So after doing a few more slow-mo of original mobs, I saw projectile's
creation rate depends on the time it has been destroyed. So to take it into calculation it was
natural to move the plasma of monster into the monster class.

Tested on the sun monster as it has a very predictable and simple pattern, now it looks almost
identical.

Also fixed sun animation, it was little borked.