dave: improve jump when near top-collision
Original game will calculate top collision during jump with the x axis applied.
However, it won't be actually applied when hitting a wall making this a bit strange.

Eventually it allows original dave to get to certain locations
(like purple platform in right side of door of secret level 4), which I want to match.

So here it is.